move setting stockstrings to frontend
