+++ src/main/deltachat/extras.ts
@@ -4,7 +4,6 @@ import setLanguage, {
   getCurrentLocaleDate,
   loadTranslations,
 } from '../load-translations'
-import { txCoreStrings } from './login'
 import { refresh as refreshMenu } from '../menu'
 
 // Extras, mainly Electron functions
@@ -17,7 +16,6 @@ export default class Extras extends SplitOut {
   }
   setLocale(locale: string) {
     setLanguage(locale)
-    this.controller.login._setCoreStrings(txCoreStrings())
     refreshMenu()
   }
 }
