+++ src/renderer/App.tsx
@@ -11,6 +11,7 @@ import { ThemeManager, ThemeContext } from './ThemeManager'
 import moment from 'moment'
 import { CrashScreen } from './components/screens/CrashScreen'
 import { runtime } from './runtime'
+import { updateCoreStrings } from './stockStrings'
 
 attachKeybindingsListener()
 
@@ -64,6 +65,7 @@ export default function App(_props: any) {
     window.static_translate = translate(localeData.messages)
     setLocaleData(localeData)
     moment.locale(localeData.locale)
+    updateCoreStrings()
   }
 
   useEffect(() => {
